add mote accurate directory detection logic for removing
