+++ internal/core/algorithm/ngt/ngt.go
@@ -26,13 +26,13 @@ import "C"
 
 import (
 	"os"
-	"path/filepath"
 	"reflect"
 	"sync"
 	"unsafe"
 
 	"github.com/vdaas/vald/internal/core/algorithm"
 	"github.com/vdaas/vald/internal/errors"
+	"github.com/vdaas/vald/internal/file"
 	"github.com/vdaas/vald/internal/log"
 )
 
@@ -241,14 +241,11 @@ func (n *ngt) loadOptions(opts ...Option) (err error) {
 }
 
 func (n *ngt) create() (err error) {
-	if fileExists(n.idxPath) {
-		log.Warnf("index path exists, will remove the directory. path: %s", n.idxPath)
-		files, err := filepath.Glob(filepath.Join(filepath.Dir(n.idxPath), "*"))
-		if err != nil {
-			return err
-		}
-		for _, file := range files {
-			err = os.RemoveAll(file)
+	files := file.ListInDir(n.idxPath)
+	if files != nil {
+		log.Warnf("index path exists, will remove the directories: %v", files)
+		for _, f := range files {
+			err = os.RemoveAll(f)
 			if err != nil {
 				return err
 			}
@@ -275,7 +272,7 @@ func (n *ngt) create() (err error) {
 }
 
 func (n *ngt) open() error {
-	if !fileExists(n.idxPath) {
+	if exists, _, _ := file.Exists(n.idxPath); !exists {
 		return errors.ErrIndexNotFound
 	}
