+++ internal/file/file_test.go
@@ -19,6 +19,7 @@ package file
 
 import (
 	"os"
+	"reflect"
 	"syscall"
 	"testing"
 
@@ -216,6 +217,180 @@ func TestOpen(t *testing.T) {
 			defer test.afterFunc(tt, test.args, got)
 
 			if err := test.checkFunc(test.want, got, err); err != nil {
+				tt.Errorf("error = %v", err)
+			}
+		})
+	}
+}
+
+func TestExists(t *testing.T) {
+	type args struct {
+		path string
+	}
+	type want struct {
+		want bool
+	}
+	type test struct {
+		name       string
+		args       args
+		want       want
+		checkFunc  func(want, bool) error
+		beforeFunc func(*testing.T, args)
+		afterFunc  func(*testing.T, args)
+	}
+	defaultCheckFunc := func(w want, got bool) error {
+		if !reflect.DeepEqual(got, w.want) {
+			return errors.Errorf("got: \"%#v\",\n\t\t\t\twant: \"%#v\"", got, w.want)
+		}
+		return nil
+	}
+
+	const (
+		baseDir      = "./utiltest"
+		testDirPath  = baseDir + "/index"
+		testFilePath = baseDir + "/ngt-meta.kvsdb"
+		testSym      = "sym"
+
+		testFailsDirPath  = baseDir + "/fails-index"
+		testFailsFilePath = baseDir + "/ngt-meta-fails.kvsdb"
+	)
+
+	defaultAfterFunc := func(t *testing.T, args args) {
+		t.Helper()
+		if err := os.RemoveAll(baseDir); err != nil {
+			t.Error(err)
+		}
+		if err := os.RemoveAll(args.path); err != nil {
+			t.Error(err)
+		}
+	}
+
+	tests := []test{
+		{
+			name: "return true when the directory exists",
+			args: args{
+				path: testDirPath,
+			},
+			beforeFunc: func(t *testing.T, args args) {
+				t.Helper()
+				if err := os.MkdirAll(args.path, 0o750); err != nil {
+					t.Fatal(err)
+				}
+			},
+			want: want{
+				want: true,
+			},
+		},
+		{
+			name: "return true when the directory exists and the type is symbolic link",
+			args: args{
+				path: testSym,
+			},
+			beforeFunc: func(t *testing.T, args args) {
+				t.Helper()
+				if err := os.MkdirAll(testDirPath, 0o750); err != nil {
+					t.Fatal(err)
+				}
+				if err := os.Symlink(testDirPath, testSym); err != nil {
+					t.Error(err)
+				}
+			},
+			want: want{
+				want: true,
+			},
+		},
+		{
+			name: "return true when the file exists",
+			args: args{
+				path: testFilePath,
+			},
+			beforeFunc: func(t *testing.T, args args) {
+				t.Helper()
+				if err := os.MkdirAll(baseDir, 0o750); err != nil {
+					t.Fatal(err)
+				}
+
+				f, err := os.Create(args.path)
+				if err != nil {
+					t.Fatal(err)
+				}
+				defer func() {
+					if err := f.Close(); err != nil {
+						t.Error(err)
+					}
+				}()
+			},
+			want: want{
+				want: true,
+			},
+		},
+		{
+			name: "return true when the file exists and the type is symbolic link",
+			args: args{
+				path: testSym,
+			},
+			beforeFunc: func(t *testing.T, args args) {
+				t.Helper()
+				if err := os.MkdirAll(baseDir, 0o750); err != nil {
+					t.Fatal(err)
+				}
+
+				f, err := os.Create(testFilePath)
+				if err != nil {
+					t.Fatal(err)
+				}
+				defer func() {
+					if err := f.Close(); err != nil {
+						t.Error(err)
+					}
+				}()
+
+				if err := os.Symlink(testFilePath, testSym); err != nil {
+					t.Error(err)
+				}
+			},
+			want: want{
+				want: true,
+			},
+		},
+		{
+			name: "return false when the directory does not exist",
+			args: args{
+				path: testFailsDirPath,
+			},
+			want: want{
+				want: false,
+			},
+		},
+		{
+			name: "return false when the file exists",
+			args: args{
+				path: testFailsFilePath,
+			},
+			want: want{
+				want: false,
+			},
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.name, func(tt *testing.T) {
+			defer goleak.VerifyNone(tt, goleak.IgnoreCurrent())
+			if test.beforeFunc != nil {
+				test.beforeFunc(tt, test.args)
+			}
+			if test.afterFunc == nil {
+				test.afterFunc = defaultAfterFunc
+			}
+			defer test.afterFunc(tt, test.args)
+
+			if test.checkFunc == nil {
+				test.checkFunc = defaultCheckFunc
+			}
+
+			// TODO we have to check more patter about file or dir
+			got, _, _ := Exists(test.args.path)
+			if err := test.checkFunc(test.want, got); err != nil {
 				tt.Errorf("error = %v", err)
 			}
 		})
